FASE 2/Evidencias Proyecto/proyecto/requirements.txt
django == 5.1
gunicorn
dj-database-url
django-admin-interface
django-crispy-forms
django-bootstrap-datepicker-plus
crispy-bootstrap5
django-rest-framework
django-anymail[mailgun]
python-dotenv
psycopg2-binary # dev, pyscopg2 if production
pillow
boto3
django-storages
pandas
openpyxl
beautifulsoup4
python-magic
rutificador
# testing packages
django-webtest
coverage
factory-boy
django-cleanup
pytest 
pytest-django